// src/universal-dispatch/domain/universal-dispatch.enums.ts
export enum UniversalServiceType {
  STANDARD_RIDE = 'STANDARD_RIDE',
  EV_SCOOTER_RIDE = 'EV_SCOOTER_RIDE',
  EV_MOTORCYCLE_RIDE = 'EV_MOTORCYCLE_RIDE',
  BICYCLE_RIDE = 'BICYCLE_RIDE',
  MINI_CAR_RIDE = 'MINI_CAR_RIDE',
  PREMIUM_RIDE = 'PREMIUM_RIDE',
  PRIVATE_DRIVER = 'PRIVATE_DRIVER',
  BUSINESS_RIDE = 'BUSINESS_RIDE',
  AIRPORT_TRANSFER = 'AIRPORT_TRANSFER',
  INTERCITY = 'INTERCITY',
  TOURISM_EVENT = 'TOURISM_EVENT',
  NIGHTLIFE = 'NIGHTLIFE',
  SENIOR_ASSISTANCE = 'SENIOR_ASSISTANCE',
  MEDICAL_PWD_RIDE = 'MEDICAL_PWD_RIDE',
  AMBULANCE_TRANSPORT = 'AMBULANCE_TRANSPORT',
  SCHOOL_RIDE = 'SCHOOL_RIDE',
  RENTAL_WITH_DRIVER = 'RENTAL_WITH_DRIVER',
  SHARED_RIDE = 'SHARED_RIDE',
  PARCEL_BIKE = 'PARCEL_BIKE',
  PARCEL_SCOOTER = 'PARCEL_SCOOTER',
  PARCEL_CAR = 'PARCEL_CAR',
  COURIER = 'COURIER',
  FOOD_DELIVERY = 'FOOD_DELIVERY',
  CARGO_VAN = 'CARGO_VAN',
  CARGO_TRUCK = 'CARGO_TRUCK',
  MULTI_STOP_ROUTE = 'MULTI_STOP_ROUTE',
  PARTNER_VEHICLE = 'PARTNER_VEHICLE',
  FLEET_CONTRACT = 'FLEET_CONTRACT',
  SCHEDULED_ROSTER = 'SCHEDULED_ROSTER',
}
// ...
export enum DispatchPolicyStatus {
  DRAFT = 'DRAFT',
  ACTIVE = 'ACTIVE',
  RETIRED = 'RETIRED',
}